Add 'guaranteed' loot list to trestype.
- Add a new loot list to TreasureType, which handles guaranteed loot
which must be given out when the mob is killed. Can specify a number
item and min/max, a non-stackable with a list of attributes to choose
from, or a non-stackable on its own.
- Move nerudite arrow drops from avar chief mob file to its trestype as
a guaranteed drop.
- Move bds from pogg king, and add it (12-20) + windroot leaf (18-30) &
inkies (8-15)as guaranteed drops.
- OPB gets 10-18 spessartine rock and 6-14 orc teeth as guaranteed drop.
- Nerudite troll gets 12-20 twinkle caps as a guaranteed drop, has its
min items dropped set to 4 (instead of i think 3 if this var isn't set)
and an increased itematt drop chance.