Add handling for DE rsc in RecordStat, sqlfixes.
- RecordStat can now use a fake kod param 'RESOURCE_GER' to handle
sending the German resource rather than English one to SQL (just use
RESOURCE_GER instead of TAG_RESOURCE in the RecordStat call).
- Added more error checking & logging to RecordStat and database.c.
- Fixed an invalid cumulative param count in C_RecordStat when calling
FreeDataNodeMemory.
- Fixed a case in BlakodDebugInfo where calling bprintf prior to kod
interpreting would cause a crash because it checked only for an invalid
class, not for a lack of interpreting entirely. Allows bprintf to
display errors prior to kod interpretation.