Add guaranteed treasure db tables.
- Change STAT_MAX to STAT_MAXSTAT in database.h because STAT_MAX is a
bit too generic and might one day collide with another constant.
- Added two new db tables (wiki_treasure_extra and wiki_treasure_magic)
to handle the plGuaranteedTreasure loot list in TreasureType. The
'extra' table handles min-max number of items (including single items
with no attributes) and the 'magic' table handles items that can take
one of a list of attributes.
- Renamed some of the variables concerning the treasure db table since
this it is more to do with TreasureType and not specifically anything to
do with Monster.
- Fixed an issue with numbitem.kod GetName when the plural name is $
(affected arrow names with number > 1).
- Fixed some treasure types with duplicate entries.