Fix up some database issues
- wiki_npc_zone no longer has npc_name_ger field, instead has npc_row
and npc_col fields to ensure NPCs are unique (e.g. keep guards on same
screen).
- wiki_npc_sellitem contains an item_color field (palette translation)
to ensure items are unique (e.g.  pants with diff colors sold by same
NPC) and so we can potentially color them in the wiki.
- Don't send data to DB for NPCs with vbSellFromInventory set (Izzio
etc) because inventory is variable and shouldn't be in the wiki.
- Removed duplicate keep guards occupying the same location in cess
castle.