Fix rotation issues zoning into guild halls.
A 105-specific change introduced angle handling with the movement
protocol which means moving also calls UserTurn. This has an edge case
where it doesn't work well - moving to a new room with an exit-specific
rotational change means that the desired angle (given by the room) won't
match the one sent by the client with the move (whichever way the user
was going prior to the room change). This is not an issue in most cases
because the client will still be facing the correct way and the incorrect
angle will be overwritten, however guild halls perform a second teleport
once a user enters the room (if the GH is owned) to make sure they are in
the foyer. This second teleport uses the incorrect angle before it gets
overwritten. The fix here is not to perform a UserTurn called by BP_MOVE
if the user is entering a new room - let the room select the angle, or
the user will keep the existing one.