Allow players to buy disabled (but accessible) spells.
Spells have two properties to mark whether they are enabled/disabled -
pbEnabled and pbAccessible. This changes spell purchasing such that only
pbAccessible matters and not pbEnabled, since the latter can be used to
temporarily disable spells which should still be available for purchase.
Can be toggled via setting in case we want to remove this live.