Fix compiler error in classvar/property checks.
When the kod compiler (blakcomp) compiles classvars and properties, they
are placed into a table datastructure (a 'symbol table') which enables
checking for duplicates (among other things). The classvar/prop name and
owning class ID are amongst the data added to the table.

Because cl/props are inherited, the symbol table for each class must
contain the cl/props of all ancestor classes, and when these are added
they have the class ID of the parent class. This is not necessarily a
bad thing, since some cl/props won't be overridden in this new class.
When creating a new cl/prop for a class, the compiler checks the table
to see if it already exists (lookup is solely by name) and if not, adds
it to the table with the class ID of this new class.

However, if it *does* exist in the table, the existing code attempts to
change the class ID to the current class, and then add it via
add_identifier(). In theory this is good, however the table_insert()
function notices a duplicate already in the table (remember, checking by
name only) and doesn't add the updated cl/prop.

Because the duplicate checks are looking for the same class ID, this
means as long as a cl/prop is defined in a parent class, the duplicate
check will never work in any child class and the cl/prop can be defined
multiple times (each subsequent definition overwriting the last).

The solution to this is to actually update the existing cl/prop class ID
in the symbol table when an override is found. When fixing this I found
quite a few duplicate cl/props which are also fixed in this commit.