Add PlayerKilledSomethingMsg to Monster.
Monster needs a PlayerKilledSomethingMsg message because when a minion
is used to kill another player, the minion itself is the stroke object
used and thus is queried for a message to display to the killer.
Monsters could override vrMinion_kill and/or PlayerKilledSomethingMsg if
they need a more descriptive message.