Fix self-cast PE spells working on others.
Self-cast only PE spells are only regulated by the client, which does
not allow the spells to be targeted to another player. However, a
modified client can send a BP_REQ_CAST with another object as the
target, and nothing server-side actually stops the cast attempt. Using
this bug, a modified client can e.g. cast invis on other players when
this shouldn't be possible.
Fixed it by explicitly checking that the target is the caster in
persench.kod. Allowed itemcast & non-players to still cast those spells
on other targets, which should be the expected behavior.