Fix bad targets in Trance's EndEnchantment
Fixes an error introduced in commit 26af9932.
In that commit, lTargets was replaced by lNewTargets in order to keep track of previous targets in lTargets. However, other references to the new target list were not updated, resulting in the original list being passed on instead.
In this commit, lTargets, again, refers to the fresh target list while the previous list of targets is passed on as lOldTargets.