Refactor kod stack dump code.
It was possible to dump the kod stack to the debug log using the
DumpStack() C call which called BlakodStackInfo() to obtain a char array
containing the full stack (including class names, message names,
parameters), however this wasn't written in a safe way. Specific stack
frames (e.g. containing a lot of parameters, esp. with long names) could
lead to a buffer overrun in some circumstances. BlakodStackInfo also
only returned part of the stack if the contents were getting too long.

This has been refactored into PrintStackToDebug which works more like
TraceInfo - a buffer of BUFFER_SIZE (10k chars, same as channel.c's
buffer) is kept, and this is flushed when it gets 'close' to full. A
full line of a stack frame is likely to be at most 3k chars (full
parameters, excessively long parameter names and contents), so the buffer
is flushed at 4k chars leaving 6k empty (so at worst the buffer could
grow to 7k during the next loop). The buffer size is also checked prior
to each snprintf to ensure a negative number is not sent to it. This
should be crash-proof but still performant.

DumpStack output can be disabled via server config option.

- Added tprintf in channel.c, which is the same as dprintf (debug log)
except with no time string. Used for printing stack dump since we don't
need time strings during a flush which might occur partway through a
stack frame line.
- Made BlakodDebugInfo in sendmsg.c (prints "filename.kod (lineno)" a
little safer.
- Adjusted GetSourceLine in class.c so it doesn't return line "0" for
the last kod instruction in a file.
- Removed an unnecessary null check in GameProcessSessionBufferUDP, as
this input is verified before calling the function (thus the check &
error message can never be hit).