Log finished quests to db, quest adjustments.
- When a user finishes a quest, log info to the database. Logs the
rewards they got, plus how long in seconds (logged in time) it took to
do the quest.
- Disable murder myster quests, these are from an old storyline and it's
going to be confusing if we just enable it like this. Players not around
prior to Akaridus being made duke likely won't even know who Kalior is.
- Display the 'quest log changed' message and a sound effect when
killing a quest monster/completing the required number of monster kills.