Fix equipped flags not updating on recv equipped item.
Obtaining an equipped item from an offer (e.g. taking armor off a
follower) causes the item to appear to keep its OF_EQUIPPED flag. The
flag is correct server-side, but flags (via ToCliObject) are sent very
early in the object handover process so that it can be drawn in the
clients inventory (while obj is still owned by the original owner). The
fix for this is just to request flags again with a post, this isn't
ideal but all other code is written assuming the current ordering of
NewHold/NewHoldObject/anything else happening so a more comprehensive
fix/refactor will have to wait.