Fix ET/reflection/original animation conflicts.
Calling SetAnimation on a player or monster in the world has the
side-effect of wiping out any animation being carried out by them as
this is being carried out in SomethingChanged calls. To fix this,
flec/ET now handle item equip/unequip separately from SomethingChanged
using the new Item calls on item equip/unequip -> item owner -> flec/et.
Also add special handling in flec/et for morph/ill form - track the
morphed status of the original player and Post a SomethingChanged if
this is modified.

Also now set up the weapon in Constructor because SomethingChanged won't
do it.

Note that I've chosen this fix specifically instead of trying to post
all the SetAnimation/SomethingChanged calls in response to the original
changing, as I think SetAnimation is generally wrong and the minion
shouldn't be trying to modify the original's properties in this way.
This does work for morph/ill form because they modify the system monster
templates which can't be viewed. This is still hacky, but doesn't risk
any visible effects.