Fix kc5/kocatan mad sci delete issues.
- Mad scientist hut tries to delete dfly queens on LastUserLeft() when
more than 4 exist in the room. However this was truly sloppily coded and
deletes everything in the room's active list after the dfly limit is
hit. This can include the NPC and also the user who just left, as they
are still in the active list at this point. If a user is present in the
room, and 5 dfly queens are added after them, the user object will be
deleted when they try to leave as the last user. Fix this by only
calling Delete on actual dragonfly queens.

KC5 (outside mad scientist hut) also has a delete issue where it caps
the number of dflies by deleting them as one enters. However it deletes
the most recent one which throws an error as it is still processing the
room change. Skip the most recent entry when capping dfly queens to
avoid this.

Note that the whole capping thing is not great as it has the potential
to delete the dfly queen a user is trying to tame (if many users are
trying to do the quest) and this could be addressed in the future.