Fix 'truth end branch' quest staying on NPC.
Parrin's truth/disciple quest has two quest templates associated with
it, and the 2nd-last part of the 1st quest will trigger the 2nd one so
the player can give variable answers (maybe a better way to do this,
idk). If the player doesn't interact with the new quest, it stays in
Parrin's quest list, visible to the user.

To remove it, there is now a QN_PRIZETYPE_DEACTIVATE_QUEST which can
deactivate a quest for a user. Will remove the branch quest when the
player completes the main quest.