Human equip GC fix, debug msg changes.
- Human mobs had their equipment set from SetEquipment called via post
from Constructor. Calling it this way is not ideal because if something
deletes the created mob during construction (e.g. mob spawning code
which fails to place it) then SetEquipment is called after deletion. To
fix this, call SetEquipment from NewOwner if the prev owner is $ and new
owner isn't, i.e. on mob placement.
- Remove a debug msg in brain.kod dealing w/ a mob in state engage when
a user enters the room. This is possible if mobs were previously
fighitng when the last user left, as they no longer stop what they are
doing when the user leaves.
- Add more info the debug msg logged when a new char creation attempt is
blocked due to invalid name. Reports the name or the length only if the
name is excessively (maliciously) long.