Track recent user move speed max vals.
To avoid rubber-banding when moving from high speed to med-low speed
(sprint wearing off, failed unleash charge) we now keep track of the
most recent move speed max values calculated for the user (the server
values, not the client-sent ones). The maximum speed used in move
calculations is then the highest value from this list/the current max.

Number tracked can be adjuisted via settings, and the list is deleted
upon logging off. Set the value in settings to 0 to turn this feature
off.