Properly handle vigor threshold changes.
NewVigor should be called when the battler has their current vigor or
vigor rest threshold changed by something (e.g. boon) so that the vigor
timer can be started if necessary. Calling NewVigor handles drawing
vigor for players already. Fixes vigor boon not causing an immediate
update in vigor rates on players, requiring them to manually do something
which triggers NewVigor to see the change.
DrawVigor in User also needs to handle a changing vigor rest threshold.
If this changes, the new value should be sent to the user.