Create a 105/112 dev team statue room.
Created a copy of the Berg throne room & allowed blinking to it.
Added dev statues in randomised positions and a book with some
notes in it. Dev statues list some contributions for each team
member.

Also fixed up some symmetry-breaking issues in the Berg throne room,
and created a slightly modified copy for the dev room.