Fix Jala priestess node travel quest.
- viQuestID is used in monster.kod for the generic 'can teach any
spells over level 2' quest, so assigning it to an NPC will gate
the spells behind that quest when HasDoneLearnQuest is called (plus
in this case, called from the Jala priestess kod file). The quest
for node travel shouldn't be assigned here, instead another easier
quest should be used.
- User has a message we can use for checking any quest
(HasCompletedQuest) which can be used for the special quest check,
and later on an easier quest can be assigned as the disciple quest
by assigning to viQuestID with no other changes required.
- Removed Lullaby from Jala priestess sell list as this spell is
too powerful outside item format without some kind of adjustment.