Fix competency/stroke/proficiency improves.
- The improve setup between stroke, proficiency and the new rogue
competency wasn't done correctly. ImproveStroke should always
call an improve message if it is called, and if they player has
at least one of the skills. This now works like before, except
proficiency improves are split between competency and proficiency
if the player has both. A further improvement might be to check
if any of the abilities are maxed so they don't waste improves,
but the original code didn't do that either (oops!).
- Fix dagger having knife proficiency instead of dagger prof.